+++ src/main/java/com/fastcampus/projectboard/service/PaginationService.java
@@ -1,0 +1,23 @@
+package com.fastcampus.projectboard.service;
+
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+@Service
+public class PaginationService {
+
+    private static final int BAR_LENGTH = 5;
+
+    public List<Integer> getPaginationBarNumber(int currentPageNumber, int totalPages){
+        int startNumber =  Math.max(currentPageNumber - (BAR_LENGTH / 2), 0);
+        int endNumber = Math.min(startNumber + BAR_LENGTH, totalPages);
+        return IntStream.range(startNumber, endNumber).boxed().collect(Collectors.toList());
+    }
+
+    public int currentBarLength(){
+        return BAR_LENGTH;
+    }
+}
