+++ src/main/resources/static/css/articles/table-header.css
@@ -1,0 +1,5 @@
+
+#article-table > thead a {
+    text-decoration: none;
+    color: black;
+}
