+++ src/main/resources/templates/articles/index.html
@@ -7,6 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
     <link href="/css/search-bar.css" rel="stylesheet">
+    <link href="/css/articles/table-header.css" rel="stylesheet">
 </head>
 
 <body>
@@ -73,10 +74,10 @@
         <table class="table" id="article-table">
             <thead>
                 <tr>
-                    <th class="title col-6">제목</th>
-                    <th class="hashtag col-2">해시태그</th>
-                    <th class="user-id col">작성자</th>
-                    <th class="created-at col">작성일</th>
+                    <th class="title col-6"><a>제목</a></th>
+                    <th class="hashtag col-2"><a>해시태그</a></th>
+                    <th class="created-id col"><a>작성자</a></th>
+                    <th class="created-at col"><a>작성일</a></th>
                 </tr>
             </thead>
             <tbody>
@@ -103,12 +104,10 @@
             </tbody>
         </table>
 
-        <nav aria-label="Page navigation example">
+        <nav id="pagination" aria-label="Page navigation example">
             <ul class="pagination justify-content-center">
                 <li class="page-item"><a class="page-link" href="#">Previous</a></li>
                 <li class="page-item"><a class="page-link" href="#">1</a></li>
-                <li class="page-item"><a class="page-link" href="#">2</a></li>
-                <li class="page-item"><a class="page-link" href="#">3</a></li>
                 <li class="page-item"><a class="page-link" href="#">Next</a></li>
             </ul>
         </nav>
