+++ src/main/resources/templates/articles/index.th.xml
@@ -4,6 +4,23 @@
     <attr sel="#footer" th:replace="footer :: footer" />
 
     <attr sel="#article-table">
+
+        <attr sel="thead/tr">
+            <attr sel="th.title/a" th:text="제목" th:href="@{/articles(page=${articles.number},
+             sort='title' + (*{articles.sort.getOrderFor('title')} != null ? (*{articles.getOrderFor('title').direction.name} != 'DESC' ? ',desc' : '' ) : '')
+            )}"/>
+        <!--    <attr sel="th.hashtag/a" th:text="해시태그" th:href="@{/articles(page=${articles.number},
+            sort='hashtag' + (*{articles.getOrderFor('hashtag')} != null ? (*{articles.getOrderFor('hashtag').direction.name} != 'DESC' ? ',desc' : '' ) : '')
+            )}"/>
+
+            <attr sel="th.created-id/a" th:text="작성자" th:href="@{/articles(page=${articles.number},
+            sort='userAccount.userId' + (*{articles.getOrderFor('userAccount.userId')} != null ? (*{articles.getOrderFor('userAccount.userId').direction.name} != 'DESC' ? ',desc' : '' ) : '')
+            )}"/>
+            <attr sel="th.created-by/a" th:text="작성일" th:href="@{/articles(page=${articles.number},
+            sort='createdAt' + (*{articles.getOrderFor('createdAt')} != null ? (*{articles.getOrderFor(createdAt).direction.name} != 'DESC' ? ',desc' : '' ) : '')
+            )}"/>-->
+        </attr>
+
         <attr sel="tbody" th:remove="all-but-first">
             <attr sel="tr[0]" th:each="article : ${articles}">
                 <attr sel="td.title/a" th:text="${article.title}" th:href="@{'/articles/'+ ${article.id}}"/>
@@ -12,5 +29,24 @@
                 <attr sel="td.created-at/time" th:datetime="${article.createdAt}" th:text="${#temporals.format(article.createdAt, 'yyyy-MM-dd')}"/>
             </attr>
         </attr>
+    </attr>
+
+    <attr sel="#pagination">
+        <attr sel="li[0]/a" th:text="Previous" th:href="@{/articles(page=${articles.number -1})}"
+        th:class="'page-link' +(${articles.number} <= 0 ? ' disabled' : '')"
+        />
+
+        <attr sel="li[1]" th:class="page-item" th:each="pageNumber : ${paginationBarNumbers}">
+            <attr sel="a"
+                th:text="${pageNumber + 1}"
+                th:href="@{/articles(page=${pageNumber})}"
+                th:class="'page-link' + (${pageNumber} == ${articles.number} ? ' disabled' : '')"
+            />
+        </attr>
+
+        <attr sel="li[2]/a" th:text="Next" th:href="@{/articles(page=${articles.number +1})}"
+              th:class="'page-link' +(${articles.number} >= ${articles.totalPages -1 } ? ' disabled' : '')"
+        />
+
     </attr>
 </thlogic>
